requirements.txt
bcrypt==4.2.1
blinker==1.8.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Flask==3.0.3
Flask-JWT==0.3.2
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
fonttools==4.39.4
greenlet==3.1.1
gunicorn==23.0.0
idna==3.3
importlib-resources==5.12.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
json2table==1.1.5
kiwisolver==1.4.4
MarkupSafe==2.1.5
matplotlib==3.7.1
numpy==2.0.2
packaging==23.1
pandas==2.2.3
pandasql==0.7.3
Pillow==9.5.0
prompt-toolkit==1.0.14
Pygments==2.16.1
PyInquirer==1.0.3
PyJWT==1.4.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2023.10.3
requests==2.28.1
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.12
wcwidth==0.2.8
Werkzeug==3.0.4
zipp==3.20.2
